Better array setting check.  We want to make sure the new array values are different before saving them.  There's no need to save if the values are the same.

+++ inc/settings/class-setting-array.php
@@ -70,11 +70,11 @@ class ButterBean_Setting_Array extends ButterBean_Setting {
 		$new_values = $this->get_posted_value();
 
 		// If there's an array of posted values, set them.
-		if ( $new_values && is_array( $new_values ) )
+		if ( $new_values && is_array( $new_values ) && $new_values !== $old_values )
 			return update_post_meta( $this->manager->post_id, $this->name, $new_values );
 
 		// If no array of posted values but we have old values, delete them.
-		else if ( $old_values )
+		else if ( $old_values && ! $new_values )
 			return delete_post_meta( $this->manager->post_id, $this->name );
 	}
 }
